python directory restructure, compiles
