+++ src/Python/color_map/color_map.h
@@ -26,6 +26,6 @@
 
 #pragma once
 
-#include <Python/open3d_pybind.h>
+#include "Python/open3d_pybind.h"
 
-void pybind_io(py::module& m);
+void pybind_color_map(py::module &m);
