+++ src/Python/geometry/geometry.h
@@ -26,6 +26,18 @@
 
 #pragma once
 
-#include <Python/open3d_pybind.h>
+#include "Python/open3d_pybind.h"
 
-void pybind_io(py::module& m);
+void pybind_geometry(py::module &m);
+
+void pybind_pointcloud(py::module &m);
+void pybind_voxelgrid(py::module &m);
+void pybind_lineset(py::module &m);
+void pybind_trianglemesh(py::module &m);
+void pybind_image(py::module &m);
+void pybind_kdtreeflann(py::module &m);
+void pybind_pointcloud_methods(py::module &m);
+void pybind_voxelgrid_methods(py::module &m);
+void pybind_lineset_methods(py::module &m);
+void pybind_trianglemesh_methods(py::module &m);
+void pybind_image_methods(py::module &m);
