+++ src/Python/geometry/geometry_trampoline.h
@@ -27,5 +27,47 @@
 #pragma once
 
 #include <Python/open3d_pybind.h>
+#include <Open3D/Geometry/Geometry.h>
+#include <Open3D/Geometry/Geometry2D.h>
+#include <Open3D/Geometry/Geometry3D.h>
 
-void pybind_io(py::module& m);
+#include "Python/geometry/geometry.h"
+
+using namespace open3d;
+
+template <class GeometryBase = Geometry>
+class PyGeometry : public GeometryBase {
+public:
+    using GeometryBase::GeometryBase;
+    void Clear() override { PYBIND11_OVERLOAD_PURE(void, GeometryBase, ); }
+    bool IsEmpty() const override {
+        PYBIND11_OVERLOAD_PURE(bool, GeometryBase, );
+    }
+};
+
+template <class Geometry3DBase = Geometry3D>
+class PyGeometry3D : public PyGeometry<Geometry3DBase> {
+public:
+    using PyGeometry<Geometry3DBase>::PyGeometry;
+    Eigen::Vector3d GetMinBound() const override {
+        PYBIND11_OVERLOAD_PURE(Eigen::Vector3d, Geometry3DBase, );
+    }
+    Eigen::Vector3d GetMaxBound() const override {
+        PYBIND11_OVERLOAD_PURE(Eigen::Vector3d, Geometry3DBase, );
+    }
+    void Transform(const Eigen::Matrix4d &transformation) override {
+        PYBIND11_OVERLOAD_PURE(void, Geometry3DBase, transformation);
+    }
+};
+
+template <class Geometry2DBase = Geometry2D>
+class PyGeometry2D : public PyGeometry<Geometry2DBase> {
+public:
+    using PyGeometry<Geometry2DBase>::PyGeometry;
+    Eigen::Vector2d GetMinBound() const override {
+        PYBIND11_OVERLOAD_PURE(Eigen::Vector2d, Geometry2DBase, );
+    }
+    Eigen::Vector2d GetMaxBound() const override {
+        PYBIND11_OVERLOAD_PURE(Eigen::Vector2d, Geometry2DBase, );
+    }
+};
