+++ src/Python/geometry/trianglemesh.cpp
@@ -24,8 +24,8 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "core.h"
-#include "core_trampoline.h"
+#include "Python/geometry/geometry_trampoline.h"
+#include "Python/geometry/geometry.h"
 
 #include <Open3D/Geometry/TriangleMesh.h>
 #include <Open3D/IO/ClassIO/TriangleMeshIO.h>
