+++ src/Python/integration/integration.h
@@ -26,33 +26,6 @@
 
 #pragma once
 
-#include <Python/py_open3d.h>
+#include "Python/open3d_pybind.h"
 
-void pybind_console(py::module &m);
-void pybind_geometry(py::module &m);
-void pybind_pointcloud(py::module &m);
-void pybind_voxelgrid(py::module &m);
-void pybind_lineset(py::module &m);
-void pybind_trianglemesh(py::module &m);
-void pybind_image(py::module &m);
-void pybind_kdtreeflann(py::module &m);
-void pybind_feature(py::module &m);
-void pybind_camera(py::module &m);
-void pybind_registration(py::module &m);
-void pybind_odometry(py::module &m);
-void pybind_global_optimization(py::module &m);
 void pybind_integration(py::module &m);
-void pybind_colormap_optimization(py::module &m);
-
-void pybind_pointcloud_methods(py::module &m);
-void pybind_voxelgrid_methods(py::module &m);
-void pybind_lineset_methods(py::module &m);
-void pybind_trianglemesh_methods(py::module &m);
-void pybind_image_methods(py::module &m);
-void pybind_feature_methods(py::module &m);
-void pybind_camera_methods(py::module &m);
-void pybind_registration_methods(py::module &m);
-void pybind_odometry_methods(py::module &m);
-void pybind_global_optimization_methods(py::module &m);
-void pybind_integration_methods(py::module &m);
-void pybind_colormap_optimization_methods(py::module &m);
