+++ src/Python/utility/utility.cpp
@@ -24,22 +24,13 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "core.h"
+#include "Python/open3d_pybind.h"
+#include "Python/utility/utility.h"
 
-#include <Open3D/Utility/Console.h>
 using namespace open3d;
 
-void pybind_console(py::module &m) {
-    py::enum_<VerbosityLevel>(m, "VerbosityLevel", py::arithmetic(),
-                              "VerbosityLevel")
-            .value("Error", VerbosityLevel::VerboseError)
-            .value("Warning", VerbosityLevel::VerboseWarning)
-            .value("Info", VerbosityLevel::VerboseInfo)
-            .value("Debug", VerbosityLevel::VerboseDebug)
-            .value("Always", VerbosityLevel::VerboseAlways)
-            .export_values();
-    m.def("set_verbosity_level", &SetVerbosityLevel,
-          "Set global verbosity level of open3d", py::arg("verbosity_level"));
-    m.def("get_verbosity_level", &GetVerbosityLevel,
-          "Get global verbosity level of open3d");
+void pybind_utility(py::module &m) {
+    py::module m_submodule = m.def_submodule("utility");
+    pybind_console(m_submodule);
+    pybind_eigen(m_submodule);
 }
