+++ src/Python/visualization/visualization.h
@@ -26,14 +26,16 @@
 
 #pragma once
 
-#include <Python/py_open3d.h>
+#include <Python/open3d_pybind.h>
+
+void pybind_visualization(py::module &m);
 
 void pybind_renderoption(py::module &m);
 void pybind_viewcontrol(py::module &m);
 void pybind_visualizer(py::module &m);
-void pybind_utility(py::module &m);
+void pybind_visualization_utility(py::module &m);
 
 void pybind_renderoption_method(py::module &m);
 void pybind_viewcontrol_method(py::module &m);
 void pybind_visualizer_method(py::module &m);
-void pybind_utility_methods(py::module &m);
+void pybind_visualization_utility_methods(py::module &m);
