+++ src/Python/visualization/visualization_trampoline.h
@@ -26,7 +26,7 @@
 
 #pragma once
 
-#include <Python/py_open3d.h>
+#include <Python/open3d_pybind.h>
 #include <Open3D/Visualization/Visualizer/Visualizer.h>
 #include <Open3D/Visualization/Visualizer/ViewControl.h>
